Added a new Kconfig ARCH_HAS_PTE_YOUNG.
The architectures that support hardware atomic update of page table
access bit ARCH_HAS_PTE_YOUNG will be selected, For other platforms it
will not get selected. arch_has_hw_pte_young() will check if
ARCH_HAS_PTE_YOUNG is enabled or not, it  will return TRUE if it
is enabled else FALSE.This change can be used in MGLRU to check if
the page table walk needed or not during aging.

Signed-off-by: Yu Zhao <yuzhao@google.com>
Signed-off-by: Donet Tom <donettom@linux.vnet.ibm.com>